Change `ExpiresIn` type in `LoginResponse` from `int` to `double`.

+++ src/Application/Dto/Response/User/LoginResponse.cs
@@ -3,5 +3,5 @@
 public record LoginResponse(
     string AccessToken,
     string RefreshToken,
-    int ExpiresIn
+    double ExpiresIn
     );
